Generate PDGs at "statement" granularity (#35)
This pull request replaces the fine-grained PDG generation with a less
granular PDG generation that doesn't break down assignments,
declarations and expressions so it's easier to read and matches
Flexeme's C# implementation.

In addition, the PRs also change the following:

- Add fine-grained tests to verify PDG nodes and edges.
- Re-implement the nameflow edge builder to only track nameflow edges
for variables, parameters, and return variable usages. The C#
implementation of the nameflow edges for literals is irrelevant and the
implementation for methods is inconsistent.

There are many opportunities for refactoring ahead of us:

- `CfgTraverser`: This class traverses the CFG using the framework
provided by `DOTCFGVisualizer`. It is not ideal because that class mixes
CFG traversal and DOT graph generation. I refactored most of the code
out of it, but we should really use a standalone CFG traverser that
doesn't assume the purpose of the traversal.
- `PdgBuilder`: The builder is starting to be bloated with the dataflow
analysis and the nameflow analysis. Ideally, the builder should be able
to take in any analysis result (format specified by an interface) and
add it to the PDG.